samples/07-caching-optimization/requirements.txt
# Sample requirements - using ToolWeaver from PyPI
toolweaver[redis]==0.1.3

# Environment variables
python-dotenv>=1.0.0
